file uploaded
Line Follower is one of the most important aspects of robotics. A Line Following Robot is an autonomous robot which is able to follow either a black line that is drawn on the surface consisting of a contrasting colour. It is designed to move automatically and follow the line. The robot uses arrays of optical sensors to identify the line, thus assisting the robot to stay on the track. The array of four sensor makes its movement precise and flexible. The robot is driven by DC gear motors to control the movement of the wheels. The aim of this project is to design a line follower to improve the sensing capability of any type of obstacle in its way and can also control speed with the help of speed regulator. This is done by using Arduino and IR sensor